+++ mayan/apps/platforms/dependencies.py
@@ -10,8 +10,5 @@ PythonDependency(
     module=__name__, name='gunicorn', version_string='==23.0.0'
 )
 PythonDependency(
-    module=__name__, name='sentry-sdk', version_string='==2.24.1'
-)
-PythonDependency(
     module=__name__, name='whitenoise', version_string='==6.9.0'
 )