+++ mayan/apps/platforms/settings.py
@@ -8,8 +8,8 @@ from mayan.settings.literals import (
 )
 
 from .literals import (
-    DEFAULT_SETTINGS_MODULE, DEFAULT_PLATFORM_CLIENT_BACKEND_ARGUMENTS,
-    DEFAULT_PLATFORM_CLIENT_BACKEND_ENABLED
+    DEFAULT_SETTINGS_MODULE, DEFAULT_PLATFORMS_CLIENT_BACKEND_ARGUMENTS,
+    DEFAULT_PLATFORMS_CLIENT_BACKEND_ENABLED
 )
 
 
@@ -18,15 +18,15 @@ setting_namespace = setting_cluster.do_namespace_add(
 )
 
 setting_client_backend_enabled = setting_namespace.do_setting_add(
-    default=DEFAULT_PLATFORM_CLIENT_BACKEND_ENABLED,
-    global_name='PLATFORM_CLIENT_BACKEND_ENABLED', help_text=_(
+    default=DEFAULT_PLATFORMS_CLIENT_BACKEND_ENABLED,
+    global_name='PLATFORMS_CLIENT_BACKEND_ENABLED', help_text=_(
         message='List of client backends to launch after startup. Use full dotted '
         'path to the client backend classes.'
     )
 )
 setting_client_backend_arguments = setting_namespace.do_setting_add(
-    default=DEFAULT_PLATFORM_CLIENT_BACKEND_ARGUMENTS,
-    global_name='PLATFORM_CLIENT_BACKEND_ARGUMENTS', help_text=_(
+    default=DEFAULT_PLATFORMS_CLIENT_BACKEND_ARGUMENTS,
+    global_name='PLATFORMS_CLIENT_BACKEND_ARGUMENTS', help_text=_(
         message='Arguments for the client backends. Use the client backend dotted '
         'path as the dictionary key for the arguments in dictionary format.'
     )